Update cabal files to comply with Hackage rules
It's a bit annoying to check all packages versions especially
when working in nix environment where they are properly set.

However it seems it would be better to follow strict rules that
are enforced by CI and set dependencies, thus we will not need
to do it once we want to publish packages on hackage.