Add reference implementation of the serializer
Reference implementation covers a highlevel interface for storing
input state to the files. At this point it has some limitations:

- single namespace is supported
- no real implementation of the merkle-tree
- all the data is kept in memory

Changing at interface to suport multiple namespaces should be pretty
straight-forward, but likely that we will add it once we have a basic
test suite.

Implementation of the merkle-tree will not change the interface and
can be pluggen in as soon as iterative merkel tree will be added
to the main branch.

For the implementation there were a few choices:

1. Builder of the chunks is implemented as an explicit functional
state machine that support commands that it interprets. There is no
common term in the ecosystem for that, and Builder was not used,
because usually Builder is a structure with Monoid instsance that
allows pure construction of the values.
Current approach does not have these properties. However it's possible
to inject that interface into differrent libraries.

2. For the reference implementation `streaming` library was used, however
it's possible to change library in the real code, current interface does
not depend on that.

During implementation some local wrappers were introduced:

- new Buffer instance for CStringLen (possibly should be a part of the mempack
  package)
- unsafeAppendToBuffer we need it bacase mempack has support only for ST
  and not direct IO, and we can't use ST becase there is no closed scope
  for working with a buffer

Introduce simple hashing

In order to check hashing introduce it.

For chunks we add a hash of the chunk content.
For entire state we are using blake28, but in the future
we will be using running merlke tree.