Add scls-cbor package
This contains the definition of 'ToCanonicalCBOR'
which can be used downstream to define the encoding
for various types to the SCLS canonical CBOR format.

We define instances for many major types, but try
to avoid cases where there may be ambiguity.

For maps and lists, we choose the indefinite length
encoding. While this is not always the most compact
encoding, it simplifies the description of what
tools should write to adhere to the SCLS
specification.