Improve namespaces support in serializer
In order to implement namespaces support we introduce
an additional API so we can get a list of namespaced groups

   Stream
      (Of namespace     <--------------- namespace
             (Stream (Of a) .. <--------  values in namespace

In the public API we allow mixing of the values from
the differrent namespaces. It's a reasonable because databases
may not support splitting namespaces and will stream values
from all namespaces at the same time.

In internal functions we do not allow that, because we want
to keep values in canonical order.

In this commit we allow only single namespace for the external-sort
implementation