Various fixes of the binary format (#60)
* Update header to not add dummy data to magic bytes

In the specification we have 'SCLS' magic bytes, but the format
used: `SCLSXYZK` where XYZK was random garbage.

This commit fixes that and creates correct header.

* Use 1 byte per network id in the hdr

Currently networkId is stored as a single byte value.
But the in the header there was a 3 bytes padding.

This commit fixes the situation so we now install a proper header.

* Introduce proper byteorder for values

Introduce Cardano.Types.ByteOrdered module in order to define
instances for the numberic types with different byteordering

And use it to introduce Hdr

* Add instances for MemPack using ByteOrder

* Remove constants hardcoding

* Restrict mempack dependency

* Remove unused extension