Increase strictness of the internal structure in the merkle-tree-iterative
Without this changes merkle tree is build lazily, and performs all the computation
only when forced. It lead to the thunks leak in the scls-file generation.

It's ok to still use lazy list and force it only to WNHF (in constuctor) even in
this form we do evaluate entire list by the construction. This is done in the finalize
as we call in during all the changes it updates only the toplevel node, and forces it
to WHNF immediately