Increase strictness of the structures
Here we increate a strictness of the structures we use, it's not
a requirement but a nice safety measure that does not cost us anything.

Using StrictData would suffice here as well, as we do not plan to use
lazyness tricks here and our code is mostly one-shot without sharing.