Strictify builder
There are two fixes that should increase strictness.

1. Force `machine` to WHNF, this is a safety measure to show the
developer that we care about the strictess (and maybe help optimizer).
We do force it by the interpretCommand anyways but this way we will
increase changes that the future developer will not something bad here.

2. In mkManifest we use strict foldl instead of sum, as by the time
of writing sum leaked thunks for some types. So it's better to be safe
than sorry.