Update namespaces to use versioning
In CIP we propose to use versioned namespaces, it would allow us
to evolve the protocol in backwards and forward compatible manner.

However in our current implementation we have not added versioning
so now we do.

Unfortunatelly now the tests do not pass because of the way how
external serializer works.