Extend information for the known Namespaces
We introduce a structure for the information about the namespaces
NamespaceInfo, that keeps an information and configuration for the
namespaces.

Currently there are only two fields, but they can be excended later:
1. namespaceSpec - Huddle specification for the namespace
2. namespaceKeySize - size of the key

We need the latter because in the CIP we propose that all the keys
in the namespaces have a statically known size. This way we can
pass this information.

We use GHC Natural because later we will need that info to pass
the values on the type level.