Loosen requirements for serializer: only keys should be Ord
Previously in a code we required entire value to have an Ord
instance. While it's possible to have such one it leads to problems
when we work with generic values of the types we do not know.

The other problem is that during the computation we should check
the equality for the keys only and do not allow two values with
the same key even if they are not equal. So the new type reflects
that properly.