Fix namespace definitions.
The definition of the UTxO and Bloks namespace relied on the ansient definition
of the generic entry, that kept both key and entry inside.
This is bad because the keys structure is defined in the .ksy definition;
it's very important as it allows to write a namespace agnostic code that
can still work with our keys:
- do search
- implement chunk generation, etc.

For this reason instead of keeping generic record we keep key information (currently size)
and value definition separately.

However with the previous approach debug tools and tests generated keys
in the values, that is incorrect. Also it made impossible to write proper tests
for generated contest.

The fix is simle, we just remove and extra part that kept a key inside a value
in the spec.