Introduce a command for checking a file
In addition to verification of the hashes this utility does internal checks:
1. It checks that number of entries in each chunk matches checksum
2. It checks the digest for each chunk
3. For all chunks with known namespaces in addition all the data is checked
against the specification.

Current approach is not the best one as currently output and verification
are done in the same place and it will be hard to extend output to a diferrent
format or not build all the results in memory.

However other commands are implemented in a similar way, so this approach
should be ok, unless it will be obvious that it does not work on the real files.