Do not include extra size in the chunks.
At this point in the chunks builder we wrap value in `Entry` that
stores size of the entry and then the entry itself. During reading
we decode value as entry. So we have a full typed roundtrip.

However once we pass a typed value in the entry for example:
  ChunkEntry — any key-value entry
  GenericChunkEntry n - key-value entry with the fixed key size
  CBOREntry n - key-value entry with the fixed key size and cbor inside

All those entries stored another prefix for the size of the entry inside.

It violates the .ksy specification, so this PR fixes the problem by removing
entra values and verifying that all tests still pass.