Fix utxo/v0 spec, buy allowing unbounded bytes in plutus_data
Plutus_data allows unterminated bytestring, even they are implented
as a unterminated list of bounded bytes that are expressed in the
spec binary representation differ for them.

They are expressed as '5F'

  5F               # bytes(*)
    58 40          # bytes(64)
       69706673..  # "ipfs.."
    42             # bytes(2)
       6634        # "f4"
    FF             # primitive(*)

Instead of '9F' for unterminated lists:

  9F              # array(*)
    58 40         # bytes(64)
       69706673.. # "ipfs.."
    01            # unsigned(1)
    FF            # primitive(*)

So we have to allow them in the spec because VBytes covers bounded_bytes
as well, we left only VBytes.

Another change is suport of typed variants in the spec as they
are used in the transactions:

tags 121..127 	alternatives 0..6, 1+1 encoding; see Section 9.    [1]
tags 1280-1400  alternatives 7..127, 1+2 encoding; see Section 9.1 [1]

[1] https://www.iana.org/go/draft-bormann-cbor-notable-tags-07