Reorganize packages
The main purpose of the change is to allow definition of the key sizes
in this repository, because downstream can't define it on it's own.

So the definitions moved to the `scls-cddl` package, where known namespaces
are defined.

However introducing that change makes scls-cddl depend on the scsl-format, and
it leads to cyclic dependencies. In order to fix it `NamespaceKeySize` was
moved to the new scls-core package. Likely we will move more types there in
the future.

To continue this process some extra dependencies between the packages were
removed. By moving Versioned from the scls-cbor to scls-core it was possible to
drop depdenency between scls-format and scls-cbor packages.

Now the structure is the following:

  - scls-core - shared types, likely downstream should not depend on that explicitly
  - scls-cbor - is needed for the libraries that define encodings (not exaclty for
                SCLS, but for any canonical CBOR types)
  - scls-cddl - is needed for libraries that define encodings for canonical ledger types
  - scls-format - is needed for the libraries or applications that performs the
                  work with the physical files.

So for example cardano-ledger will depend only on scls-cbor and scls-cddl as it
defines types for the ledger state.

ledger-state application will depend on cardano-ledger and scls-format.

Hypothetical package that utilise scls container will depend only on scls-cbor or scls-format
(if works with container).